Transfer over tests, refine testing infrastructure
