+++ Sources/BeamedRhythm/RhythmSpelling.TieState.swift
@@ -6,22 +6,22 @@
 //
 //
 
-extension RhythmSpelling {
-    
-    /// Whether a tie is needed to start, stop, or not exist at all for a given
-    /// `RhythmSpelling.Context`.
-    public enum TieState {
-
-        /// No ties needed.
-        case none
-        
-        /// Start tie.
-        case start
-        
-        /// Stop tie.
-        case stop
-        
-        /// Continue a tie by stopping the previous and starting a new one.
-        case maintain
-    }
-}
+//extension RhythmSpelling {
+//    
+//    /// Whether a tie is needed to start, stop, or not exist at all for a given
+//    /// `RhythmSpelling.Context`.
+//    public enum TieState {
+//
+//        /// No ties needed.
+//        case none
+//        
+//        /// Start tie.
+//        case start
+//        
+//        /// Stop tie.
+//        case stop
+//        
+//        /// Continue a tie by stopping the previous and starting a new one.
+//        case maintain
+//    }
+//}
